opt: SEQUOIADBMAINSTREAM-4359 实现通用的 LightJob 机制
git-svn-id: http://192.168.20.11/sequoiadb/trunk@40461 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/util/utilMemBlockPool.cpp
@@ -80,7 +80,7 @@ namespace engine
                               UTIL_MEM_A_BLOCK_SUBPOOL_NUM ;
       UINT32 bigBlockSize = UTIL_MEM_A_BIG_BLOCK_SIZE *
                             UTIL_MEM_A_BLOCK_SUBPOOL_NUM ;
-      UINT32 aBlockMaxSize = 0 ;
+      UINT64 aBlockMaxSize = 0 ;
 
       if ( 0 != _maxSize )
       {
